pyproject.toml
[project]
name = "web-wg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=44.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.2",
    "uvicorn>=0.34.0",
]
